+++ testing/test-utils/src/main/kotlin/net/corda/testing/node/MockAttachmentStorage.kt
@@ -4,7 +4,10 @@ import net.corda.core.contracts.Attachment
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.sha256
 import net.corda.core.internal.AbstractAttachment
+import net.corda.core.node.services.AttachmentId
 import net.corda.core.node.services.AttachmentStorage
+import net.corda.core.node.services.vault.AttachmentQueryCriteria
+import net.corda.core.node.services.vault.AttachmentSort
 import net.corda.core.serialization.SingletonSerializeAsToken
 import java.io.ByteArrayOutputStream
 import java.io.InputStream
@@ -12,16 +15,8 @@ import java.util.HashMap
 import java.util.jar.JarInputStream
 
 class MockAttachmentStorage : AttachmentStorage, SingletonSerializeAsToken() {
-    val files = HashMap<SecureHash, ByteArray>()
 
-    override fun openAttachment(id: SecureHash): Attachment? {
-        val f = files[id] ?: return null
-        return object : AbstractAttachment({ f }) {
-            override val id = id
-        }
-    }
-
-    override fun importAttachment(jar: InputStream): SecureHash {
+    override fun importAttachment(jar: InputStream): AttachmentId {
         // JIS makes read()/readBytes() return bytes of the current file, but we want to hash the entire container here.
         require(jar !is JarInputStream)
 
@@ -36,5 +31,22 @@ class MockAttachmentStorage : AttachmentStorage, SingletonSerializeAsToken() {
             files[sha256] = bytes
         }
         return sha256
+    }
+
+    override fun importAttachment(jar: InputStream, uploader: String, filename: String): AttachmentId {
+        return importAttachment(jar)
+    }
+
+    val files = HashMap<SecureHash, ByteArray>()
+
+    override fun openAttachment(id: SecureHash): Attachment? {
+        val f = files[id] ?: return null
+        return object : AbstractAttachment({ f }) {
+            override val id = id
+        }
+    }
+
+    override fun queryAttachments(criteria: AttachmentQueryCriteria, sorting: AttachmentSort?): List<AttachmentId> {
+        throw NotImplementedError("Querying for attachments not implemented")
     }
 }